added functions to save wallet data
store.go has all methods concerning saving, wallet info saved in gob encoding
